Move Zaptec API functionality to separate zaptec/ directory. This is done to split the responsibility between the API facing parts and the HA facing parts.
