+++ custom_components/zaptec/services.py
@@ -12,8 +12,8 @@ from homeassistant.helpers import device_registry as dr, entity_registry as er
 import homeassistant.helpers.config_validation as cv
 import voluptuous as vol
 
-from .api import Charger, Installation
 from .const import DOMAIN
+from .zaptec import Charger, Installation
 
 if TYPE_CHECKING:
     from .coordinator import ZaptecUpdateCoordinator
