+++ custom_components/zaptec/update.py
@@ -17,9 +17,9 @@ from homeassistant.exceptions import HomeAssistantError
 from homeassistant.helpers.entity import EntityDescription
 from homeassistant.helpers.entity_platform import AddEntitiesCallback
 
-from .api import Charger
 from .entity import ZaptecBaseEntity
 from .manager import ZaptecConfigEntry
+from .zaptec import Charger
 
 _LOGGER = logging.getLogger(__name__)
