+++ custom_components/zaptec/zaptec/misc.py
