+++ custom_components/zaptec/zaptec/validate.py
