+++ custom_components/zaptec/zaptec/zconst.py
